feat(scheduler): agrega modelo de dominio y persistencia para farmacias en turno
CAMBIOS:
- Introduce entidades JPA (Pharmacy, PharmacyOnDuty, PkEntityBase) para mapear tablas de farmacias y sus turnos
- Agrega repositorios Spring Data JPA para operaciones CRUD sobre Pharmacy y PharmacyOnDuty
- Implementa PharmaManager como servicio transaccional para orquestar persistencia de farmacias con validaciones de duplicados
- Crea RefreshTask como tarea programada que obtiene datos de MINSAL y los persiste cada hora
- Agrega utilidades PhoneNumberUtils y TextUtils para normalización de números telefónicos y texto con locale es-CL
- Migra MinsalService de RestTemplate a RestClient para simplificar configuración de headers y manejo de requests
- Cambia tipo de storeId en FarmaciaTurno de Long a Integer para alinearse con el modelo de dominio
- Modifica duty_date en tabla pharmacies_on_duty de timestamp a date para reflejar que solo se almacena la fecha
- Agrega commons-text y commons-collections4 como dependencias para utilidades de texto y colecciones
- Configura datasource PostgreSQL, pool de hilos para @Scheduled y período de shutdown en application.properties
- Proporciona application.properties separado para tests con H2 en memoria

ANALISIS:
- Archivos afectados: db/01-model.sql, scheduler/pom.xml, MinsalService.java, FarmaciaTurno.java, application.properties, + 9 archivos nuevos
- Tipo de cambio: feature
- Impacto: aditivo (agrega nuevo modelo y persistencia; migración de RestTemplate)

+++ scheduler/src/main/java/cl/sebastian/cm/scheduler/domain/data/FarmaciaTurno.java
@@ -32,7 +32,7 @@ public class FarmaciaTurno extends Seba {
 
     @Schema(description = "Identificador único del local de farmacia", example = "12345")
     @JsonProperty("local_id")
-    private Long storeId;
+    private Integer storeId;
 
     @Schema(description = "Clave foránea de la región (según codificación externa)", example = "8")
     @JsonProperty("fk_region")
@@ -110,7 +110,7 @@ public class FarmaciaTurno extends Seba {
      *
      * @return ID de la farmacia
      */
-    public Long getStoreId() {
+    public Integer getStoreId() {
         return storeId;
     }
 
@@ -119,7 +119,7 @@ public class FarmaciaTurno extends Seba {
      *
      * @param storeId ID de la farmacia
      */
-    public void setStoreId(Long storeId) {
+    public void setStoreId(Integer storeId) {
         this.storeId = storeId;
     }
